requirements.txt
appium-python-client
Pillow
python-dotenv
requests
sqlite
lxml
opencv-python
flask
schedule